Added BVH, switched to `nalgebra` rays/vectors, dropped motion blur support
Now uses a bounding volume hierarchy from the `bvh` crate to traverse
the scene and find ray-shape intersections faster. As a result of `bvh`
using `nalgebra` for its vector and ray types, switched to `nalgebra`.
This necessitated dropping support for motion blur, as the `ray` struct
required a time component for the previous implementation, which
`nalgebra` does not support.
Also, renamed/refactored `Hittable` and `HitRecord` to `Shape`
`Intersection` for clarity.